# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "description": "Returns a short-lived access token in the body and sets the rotating refresh token as an HttpOnly cookie. The refresh token never appears in a response body.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/LoginResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The account is disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many sign-in attempts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "description": "Rotates the refresh token: the presented one is revoked as the replacement is issued. Presenting an already-used token revokes the whole session family.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/LoginResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "The refresh token is missing, expired, or already used.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "rb_refresh": []
          }
        ],
        "summary": "Exchange the refresh cookie for a new access token",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/session": {
      "get": {
        "description": "Re-reads the user from the database, so a role change or a disabled account takes effect at once.",
        "operationId": "AuthController_session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/SessionResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "VALIDATION_ERROR"
                        },
                        "message": {
                          "type": "string",
                          "example": "Request validation failed."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true,
                          "description": "Present only when the error carries structured context."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the current user",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "React Boilerplate API",
    "description": "Backend for the React boilerplate. Every response uses the shared envelope: `{ success: true, data }` on success and `{ success: false, error: { code, message, details? } }` on failure. HTTP status describes the transport outcome; `error.code` describes the domain outcome.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "rb_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rb_refresh"
      }
    },
    "schemas": {
      "AuthUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a3f1c2e4-..."
          },
          "email": {
            "type": "string",
            "example": "demo@example.com"
          },
          "name": {
            "type": "string",
            "example": "Demo Maker"
          },
          "permissions": {
            "example": [
              "dashboard:read",
              "user:read"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "permissions"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "demo@example.com"
          },
          "password": {
            "type": "string",
            "example": "correct horse battery staple",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          }
        },
        "required": [
          "user"
        ]
      },
...
    }
  }
}